refactor: cart validation 분리 & 쿼리최적화
